feat/LJJ(#112): useObserver 구현해 무한 스크롤 부분 모듈화
